Fixed issue with cabal file. I had deleted the license file provided by cabal since git provides one. HOWEVER, cabal requires it's license file to be there in order to compile. I pointed the cabal config to the git license file.
There was a bug where if an invalid tag was found, the parser would exit and return a "successful" result of whatever it parsed up to that point. Changed the plain text parser to do manytill a tag is found, not many till a delimiter is found. Now it treats "bad tags" as plain text and doesn't parse them.